+++ src/arbitrary/_internals/CloneArbitrary.ts
@@ -1,13 +1,11 @@
+import { NextArbitrary } from '../../check/arbitrary/definition/NextArbitrary';
+import { NextValue } from '../../check/arbitrary/definition/NextValue';
+import { cloneMethod } from '../../check/symbols';
 import { Random } from '../../random/generator/Random';
 import { Stream } from '../../stream/Stream';
-import { cloneMethod } from '../symbols';
-import { Arbitrary } from './definition/Arbitrary';
-import { convertFromNext, convertToNext } from './definition/Converters';
-import { NextArbitrary } from './definition/NextArbitrary';
-import { NextValue } from './definition/NextValue';
 
 /** @internal */
-class CloneArbitrary<T> extends NextArbitrary<T[]> {
+export class CloneArbitrary<T> extends NextArbitrary<T[]> {
   constructor(readonly arb: NextArbitrary<T>, readonly numValues: number) {
     super();
   }
@@ -86,35 +84,3 @@ class CloneArbitrary<T> extends NextArbitrary<T[]> {
     return new NextValue(vs, contexts);
   }
 }
-
-/**
- * Type of the value produced by {@link clone}
- * @remarks Since 2.5.0
- * @public
- */
-export type CloneValue<T, N extends number> = N extends 0
-  ? []
-  : N extends 1
-  ? [T]
-  : N extends 2
-  ? [T, T]
-  : N extends 3
-  ? [T, T, T]
-  : N extends 4
-  ? [T, T, T, T]
-  : T[];
-
-/**
- * Clone the values generated by `arb` in order to produce fully equal values (might not be equal in terms of === or ==)
- *
- * @param arb - Source arbitrary
- * @param numValues - Number of values to produce
- *
- * @remarks Since 2.5.0
- * @public
- */
-function clone<T, N extends number>(arb: Arbitrary<T>, numValues: N): Arbitrary<CloneValue<T, N>>;
-function clone<T>(arb: Arbitrary<T>, numValues: number): Arbitrary<T[]> {
-  return convertFromNext(new CloneArbitrary(convertToNext(arb), numValues));
-}
-export { clone };
