+++ src/fast-check-default.ts
@@ -32,8 +32,8 @@ import { constant } from './arbitrary/constant';
 import { constantFrom } from './arbitrary/constantFrom';
 import { context, ContextValue } from './arbitrary/context';
 import { date } from './check/arbitrary/DateArbitrary';
-import { clone, CloneValue } from './check/arbitrary/CloneArbitrary';
-import { dedup, DedupValue } from './check/arbitrary/DedupArbitrary';
+import { clone, CloneValue } from './arbitrary/clone';
+import { dedup, DedupValue } from './arbitrary/dedup';
 import { Arbitrary } from './check/arbitrary/definition/Arbitrary';
 import { Shrinkable } from './check/arbitrary/definition/Shrinkable';
 import { dictionary } from './check/arbitrary/DictionaryArbitrary';
